+++ tests/com/kaylerrenslow/armaDialogCreator/arma/header/largeTest/HeaderParserLargeTest.java
@@ -15,7 +15,7 @@ public class HeaderParserLargeTest {
 
 	@Test
 	public void parseHeaderTest1() throws Exception {
-		HeaderFile headerFile = HeaderParser.parse(HeaderTestUtil.getFile("largeTest/test1Files/largeTestRoot.h"));
+		HeaderFile headerFile = HeaderParser.parse(HeaderTestUtil.getFile("largeTest/test1Files/largeTestRoot.h"), HeaderTestUtil.getTemporaryResultsFile());
 
 		HeaderClass expected = hClass("-root class", null,
 				assign("author", w("K-Town")),
@@ -68,7 +68,7 @@ public class HeaderParserLargeTest {
 
 	@Test
 	public void parseHeaderTest2() throws Exception {
-		HeaderFile headerFile = HeaderParser.parse(HeaderTestUtil.getFile("largeTest/test2Files/largeTest2.h"));
+		HeaderFile headerFile = HeaderParser.parse(HeaderTestUtil.getFile("largeTest/test2Files/largeTest2.h"), HeaderTestUtil.getTemporaryResultsFile());
 
 		HeaderClass expected = hClass("-root class", null,
 				hClass("DevStuff", null,
@@ -85,7 +85,7 @@ public class HeaderParserLargeTest {
 
 	@Test
 	public void parseHeaderTest3() throws Exception {
-		HeaderFile headerFile = HeaderParser.parse(HeaderTestUtil.getFile("largeTest/test3Files/largeTest3.h"));
+		HeaderFile headerFile = HeaderParser.parse(HeaderTestUtil.getFile("largeTest/test3Files/largeTest3.h"), HeaderTestUtil.getTemporaryResultsFile());
 
 		HeaderClass hClassOne = hClass(
 				"Sound_Man_one",