+++ tests/com/kaylerrenslow/armaDialogCreator/control/ControlClassInheritanceTests.java
@@ -1,0 +1,14 @@
+package com.kaylerrenslow.armaDialogCreator.control;
+
+import org.junit.Test;
+
+/**
+ Tests for {@link ControlClass} that tests inheritance ({@link ControlClass#extendControlClass(ControlClass)}) in many different ways
+
+ @author Kayler
+ @since 05/30/2017 */
+public class ControlClassInheritanceTests {
+	@Test
+	public void overrideProperty() throws Exception {
+	}
+}