+++ tests/com/kaylerrenslow/armaDialogCreator/control/ControlPropertyTest.java
@@ -113,11 +113,20 @@ public class ControlPropertyTest {
 	}
 
 	@Test
-	public void deepCopy() throws Exception {
+	public void deepCopy_succeed() throws Exception {
 		AColor color = new AColor(0, 0, 0, 1);
 		ControlProperty p = new ControlProperty(ControlPropertyLookup.IDC, color);
 		ControlProperty pCopy = p.deepCopy();
 		assertEquals(true, p != pCopy && p.equals(pCopy));
+	}
+
+	@Test
+	public void deepCopy_fail() throws Exception {
+		AColor color = new AColor(0, 0, 0, 1);
+		ControlProperty p = new ControlProperty(ControlPropertyLookup.IDC, color);
+		ControlProperty pCopy = p.deepCopy();
+		pCopy.setValue(new SVString("Why so serious?"));
+		assertEquals(false, p.equals(pCopy));
 	}
 
 	@Test